将AirTest_MobileAuto框架移动到pypi仓库
