+++ bin/run.sh
@@ -6,4 +6,4 @@ para="$1 1"
 else
 para=""
 fi
-python3 -u object.py $para  2>&1 | tee result.$1
+python3 -u wzry.py $para  2>&1 | tee result.$1
