--- bin/run3.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-echo ckill python
-echo ckill object
-python3 -u object.py -3 2>&1 | tee result.3
-
