+++ run.bat
@@ -1,3 +1,3 @@
 @echo off
 chcp 65001
-%USERPROFILE%\AppData\Local\anaconda3\python.exe object.py  %1 %2 %3
+%USERPROFILE%\AppData\Local\anaconda3\python.exe wzry.py  %*
