+++ run2.bat
@@ -1,1 +1,1 @@
-%USERPROFILE%\AppData\Local\anaconda3\python.exe object.py -2
+.\run.bat -2
